removed np float 64 from the coding, absolutely free from external resources now

+++ src/emission.py
@@ -1,7 +1,6 @@
 import string
 import time
 import toolbox as tool
-import numpy as np
 class emission(object):
     def __init__(self):
         self.matrix = {}
@@ -66,7 +65,7 @@ class emission(object):
         else:
             raise RuntimeError("Tag '"+tag+"' not found")
 
-        return np.float64(numer*1.0)/np.float64(denom+1.0)
+        return numer*1.0/(denom+1.0)
     # predict the labels of the input file
     def predict(self, infile, outfile, p=True):
         # start=time.clock()
